backend/app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi[all]
azure.ai.formrecognizer
azure.core
cryptography
azure-storage-blob
azure-identity
requests
pandas
numpy
